requirements.txt
Django>=3.2.0,<4.1.0
djangorestframework>=3.12.0,<3.14.0
django-filter==21.1
djangorestframework-simplejwt==5.1.0
drf-yasg2>=1.19.0
flake8>=4.0.0,<4.1.0
